@@ -13,8 +13,10 @@ namespace Liip\ImagineBundle\Binary\Loader;
 
 use League\Flysystem\FilesystemInterface;
 use Liip\ImagineBundle\Exception\Binary\Loader\NotLoadableException;
+use Liip\ImagineBundle\Exception\InvalidArgumentException;
 use Liip\ImagineBundle\Model\Binary;
-use Symfony\Component\HttpFoundation\File\MimeType\ExtensionGuesserInterface;
+use Symfony\Component\HttpFoundation\File\MimeType\ExtensionGuesserInterface as DeprecatedExtensionGuesserInterface;
+use Symfony\Component\Mime\MimeTypesInterface;
 
 class FlysystemLoader implements LoaderInterface
 {
@@ -24,14 +26,22 @@ class FlysystemLoader implements LoaderInterface
     protected $filesystem;
 
     /**
-     * @var ExtensionGuesserInterface
+     * @var MimeTypesInterface|DeprecatedExtensionGuesserInterface
      */
     protected $extensionGuesser;
 
     public function __construct(
-        ExtensionGuesserInterface $extensionGuesser,
+        $extensionGuesser,
         FilesystemInterface $filesystem)
     {
+        if (!$extensionGuesser instanceof MimeTypesInterface && !$extensionGuesser instanceof DeprecatedExtensionGuesserInterface) {
+            throw new InvalidArgumentException('$extensionGuesser must be an instance of Symfony\Component\Mime\MimeTypesInterface or Symfony\Component\HttpFoundation\File\MimeType\ExtensionGuesserInterface');
+        }
+
+        if (interface_exists(MimeTypesInterface::class) && $extensionGuesser instanceof DeprecatedExtensionGuesserInterface) {
+            @trigger_error(sprintf('Passing a %s to "%s()" is deprecated since Symfony 4.3, pass a "%s" instead.', DeprecatedExtensionGuesserInterface::class, __METHOD__, MimeTypesInterface::class), E_USER_DEPRECATED);
+        }
+
         $this->extensionGuesser = $extensionGuesser;
         $this->filesystem = $filesystem;
     }
@@ -47,10 +57,25 @@ class FlysystemLoader implements LoaderInterface
 
         $mimeType = $this->filesystem->getMimetype($path);
 
+        $extension = $this->getExtension($mimeType);
+
         return new Binary(
             $this->filesystem->read($path),
             $mimeType,
-            $this->extensionGuesser->guess($mimeType)
+            $extension
         );
+    }
+
+    private function getExtension(?string $mimeType): ?string
+    {
+        if ($this->extensionGuesser instanceof DeprecatedExtensionGuesserInterface) {
+            return $this->extensionGuesser->guess($mimeType);
+        }
+
+        if (null === $mimeType) {
+            return null;
+        }
+
+        return $this->extensionGuesser->getExtensions($mimeType)[0] ?? null;
     }
 }
