@@ -23,12 +23,13 @@ class FileSystemLocator implements LocatorInterface
 
     /**
      * @param string[] $roots
+     * @param bool     $allowUnresolvable
      */
-    public function __construct(array $roots = [])
+    public function __construct(array $roots = [], bool $allowUnresolvable = false)
     {
-        $this->roots = array_map(function (string $root): string {
-            return $this->sanitizeRootPath($root);
-        }, $roots);
+        $this->roots = array_filter(array_map(function (string $root) use ($allowUnresolvable): ?string {
+            return $this->sanitizeRootPath($root, $allowUnresolvable);
+        }, $roots));
     }
 
     /**
@@ -104,15 +105,20 @@ class FileSystemLocator implements LocatorInterface
 
     /**
      * @param string $path
+     * @param bool   $allowUnresolvable
      *
      * @throws InvalidArgumentException
      *
-     * @return string
+     * @return string|null
      */
-    private function sanitizeRootPath(string $path): string
+    private function sanitizeRootPath(string $path, bool $allowUnresolvable): ?string
     {
         if (!empty($path) && false !== $real = realpath($path)) {
             return $real;
+        }
+
+        if ($allowUnresolvable) {
+            return null;
         }
 
         throw new InvalidArgumentException(sprintf('Root image path not resolvable "%s"', $path));
@@ -131,7 +137,7 @@ class FileSystemLocator implements LocatorInterface
             return 0 === mb_strpos($path, $root);
         });
 
-        if (0 === count($roots)) {
+        if (0 === \count($roots)) {
             throw new NotLoadableException(
                 sprintf('Source image invalid "%s" as it is outside of the defined root path(s) "%s"', $path, implode(':', $this->roots))
             );
