@@ -212,7 +212,7 @@ class CacheResolver implements ResolverInterface
         if ($this->cache->contains($indexKey)) {
             $index = (array) $this->cache->fetch($indexKey);
 
-            if (!in_array($cacheKey, $index, true)) {
+            if (!\in_array($cacheKey, $index, true)) {
                 $index[] = $cacheKey;
             }
         } else {
@@ -236,7 +236,7 @@ class CacheResolver implements ResolverInterface
     {
         $resolver->setDefaults([
             'global_prefix' => 'liip_imagine.resolver_cache',
-            'prefix' => get_class($this->resolver),
+            'prefix' => \get_class($this->resolver),
             'index_key' => 'index',
         ]);
 
