@@ -12,7 +12,7 @@
 namespace Liip\ImagineBundle\Imagine\Cache\Resolver;
 
 use League\Flysystem\AdapterInterface;
-use League\Flysystem\Filesystem;
+use League\Flysystem\FilesystemInterface;
 use Liip\ImagineBundle\Binary\BinaryInterface;
 use Liip\ImagineBundle\Exception\Imagine\Cache\Resolver\NotResolvableException;
 use Symfony\Component\Routing\RequestContext;
@@ -20,7 +20,7 @@ use Symfony\Component\Routing\RequestContext;
 class FlysystemResolver implements ResolverInterface
 {
     /**
-     * @var Filesystem
+     * @var FilesystemInterface
      */
     protected $flysystem;
 
@@ -56,14 +56,14 @@ class FlysystemResolver implements ResolverInterface
     /**
      * FlysystemResolver constructor.
      *
-     * @param Filesystem     $flysystem
-     * @param RequestContext $requestContext
-     * @param string         $rootUrl
-     * @param string         $cachePrefix
-     * @param string         $visibility
+     * @param FilesystemInterface $flysystem
+     * @param RequestContext      $requestContext
+     * @param string              $rootUrl
+     * @param string              $cachePrefix
+     * @param string              $visibility
      */
     public function __construct(
-        Filesystem $flysystem,
+        FilesystemInterface $flysystem,
         RequestContext $requestContext,
         $rootUrl,
         $cachePrefix = 'media/cache',
@@ -105,8 +105,8 @@ class FlysystemResolver implements ResolverInterface
     {
         return sprintf(
             '%s/%s',
-            $this->webRoot,
-            $this->getFileUrl($path, $filter)
+            rtrim($this->webRoot, '/'),
+            ltrim($this->getFileUrl($path, $filter), '/')
         );
     }
 
@@ -122,7 +122,7 @@ class FlysystemResolver implements ResolverInterface
         $this->flysystem->put(
             $this->getFilePath($path, $filter),
             $binary->getContent(),
-            ['visibility' => $this->visibility]
+            ['visibility' => $this->visibility, 'mimetype' => $binary->getMimeType()]
         );
     }
 
