@@ -12,6 +12,7 @@
 namespace Liip\ImagineBundle\Imagine\Cache\Resolver;
 
 use Liip\ImagineBundle\Binary\BinaryInterface;
+use Liip\ImagineBundle\Imagine\Cache\Helper\PathHelper;
 use Symfony\Component\Routing\RequestContext;
 
 class NoCacheWebPathResolver implements ResolverInterface
@@ -42,10 +43,19 @@ class NoCacheWebPathResolver implements ResolverInterface
      */
     public function resolve($path, $filter)
     {
-        return sprintf('%s://%s/%s',
+        $port = '';
+        if ('https' === $this->requestContext->getScheme() && 443 !== $this->requestContext->getHttpsPort()) {
+            $port = ":{$this->requestContext->getHttpsPort()}";
+        }
+        if ('http' === $this->requestContext->getScheme() && 80 !== $this->requestContext->getHttpPort()) {
+            $port = ":{$this->requestContext->getHttpPort()}";
+        }
+
+        return sprintf('%s://%s%s/%s',
             $this->requestContext->getScheme(),
             $this->requestContext->getHost(),
-            ltrim($path, '/')
+            $port,
+            ltrim(PathHelper::filePathToUrlPath($path), '/')
         );
     }
 
