@@ -14,9 +14,11 @@ namespace Liip\ImagineBundle\Imagine\Data;
 use Liip\ImagineBundle\Binary\BinaryInterface;
 use Liip\ImagineBundle\Binary\Loader\LoaderInterface;
 use Liip\ImagineBundle\Binary\MimeTypeGuesserInterface;
+use Liip\ImagineBundle\Exception\InvalidArgumentException;
 use Liip\ImagineBundle\Imagine\Filter\FilterConfiguration;
 use Liip\ImagineBundle\Model\Binary;
-use Symfony\Component\HttpFoundation\File\MimeType\ExtensionGuesserInterface;
+use Symfony\Component\HttpFoundation\File\MimeType\ExtensionGuesserInterface as DeprecatedExtensionGuesserInterface;
+use Symfony\Component\Mime\MimeTypesInterface;
 
 class DataManager
 {
@@ -26,7 +28,7 @@ class DataManager
     protected $mimeTypeGuesser;
 
     /**
-     * @var ExtensionGuesserInterface
+     * @var DeprecatedExtensionGuesserInterface|MimeTypesInterface
      */
     protected $extensionGuesser;
 
@@ -51,19 +53,27 @@ class DataManager
     protected $loaders = [];
 
     /**
-     * @param MimeTypeGuesserInterface  $mimeTypeGuesser
-     * @param ExtensionGuesserInterface $extensionGuesser
-     * @param FilterConfiguration       $filterConfig
-     * @param string                    $defaultLoader
-     * @param string                    $globalDefaultImage
+     * @param MimeTypeGuesserInterface                               $mimeTypeGuesser
+     * @param DeprecatedExtensionGuesserInterface|MimeTypesInterface $extensionGuesser
+     * @param FilterConfiguration                                    $filterConfig
+     * @param string                                                 $defaultLoader
+     * @param string                                                 $globalDefaultImage
      */
     public function __construct(
         MimeTypeGuesserInterface $mimeTypeGuesser,
-        ExtensionGuesserInterface $extensionGuesser,
+        $extensionGuesser,
         FilterConfiguration $filterConfig,
         $defaultLoader = null,
         $globalDefaultImage = null
     ) {
+        if (!$extensionGuesser instanceof MimeTypesInterface && !$extensionGuesser instanceof DeprecatedExtensionGuesserInterface) {
+            throw new InvalidArgumentException('$extensionGuesser must be an instance of Symfony\Component\Mime\MimeTypesInterface or Symfony\Component\HttpFoundation\File\MimeType\ExtensionGuesserInterface');
+        }
+
+        if (interface_exists(MimeTypesInterface::class) && $extensionGuesser instanceof DeprecatedExtensionGuesserInterface) {
+            @trigger_error(sprintf('Passing a %s to "%s()" is deprecated since Symfony 4.3, pass a "%s" instead.', DeprecatedExtensionGuesserInterface::class, __METHOD__, MimeTypesInterface::class), E_USER_DEPRECATED);
+        }
+
         $this->mimeTypeGuesser = $mimeTypeGuesser;
         $this->filterConfig = $filterConfig;
         $this->defaultLoader = $defaultLoader;
@@ -126,10 +136,11 @@ class DataManager
         if (!$binary instanceof BinaryInterface) {
             $mimeType = $this->mimeTypeGuesser->guess($binary);
 
+            $extension = $this->getExtension($mimeType);
             $binary = new Binary(
                 $binary,
                 $mimeType,
-                $this->extensionGuesser->guess($mimeType)
+                $extension
             );
         }
 
@@ -163,5 +174,18 @@ class DataManager
         }
 
         return $defaultImage;
+    }
+
+    private function getExtension(?string $mimeType): ?string
+    {
+        if ($this->extensionGuesser instanceof DeprecatedExtensionGuesserInterface) {
+            return $this->extensionGuesser->guess($mimeType);
+        }
+
+        if (null === $mimeType) {
+            return null;
+        }
+
+        return $this->extensionGuesser->getExtensions($mimeType)[0] ?? null;
     }
 }
