@@ -12,31 +12,25 @@
 namespace Liip\ImagineBundle\Imagine\Filter\PostProcessor;
 
 use Liip\ImagineBundle\Binary\BinaryInterface;
-use Liip\ImagineBundle\Binary\FileBinaryInterface;
+use Liip\ImagineBundle\Exception\Imagine\Filter\PostProcessor\InvalidOptionException;
 use Liip\ImagineBundle\Model\Binary;
 use Symfony\Component\Process\Exception\ProcessFailedException;
-use Symfony\Component\Process\Process;
 
-class JpegOptimPostProcessor implements PostProcessorInterface
+class JpegOptimPostProcessor extends AbstractPostProcessor
 {
-    /**
-     * @var string Path to jpegoptim binary
-     */
-    protected $jpegoptimBin;
-
     /**
      * If set --strip-all will be passed to jpegoptim.
      *
      * @var bool
      */
-    protected $stripAll;
+    protected $strip;
 
     /**
      * If set, --max=$value will be passed to jpegoptim.
      *
      * @var int
      */
-    protected $max;
+    protected $quality;
 
     /**
      * If set to true --all-progressive will be passed to jpegoptim, otherwise --all-normal will be passed.
@@ -46,129 +40,135 @@ class JpegOptimPostProcessor implements PostProcessorInterface
     protected $progressive;
 
     /**
-     * Directory where temporary file will be written.
-     *
-     * @var string
+     * @param string $executablePath    Path to the jpegoptim binary
+     * @param bool   $strip             Strip all markers from output
+     * @param int    $quality           Set maximum image quality factor
+     * @param bool   $progressive       Force output to be progressive
+     * @param string $temporaryRootPath Directory where temporary file will be written
      */
-    protected $tempDir;
+    public function __construct($executablePath = '/usr/bin/jpegoptim', $strip = true, $quality = null, $progressive = true, $temporaryRootPath = null)
+    {
+        parent::__construct($executablePath, $temporaryRootPath);
 
-    /**
-     * Constructor.
-     *
-     * @param string $jpegoptimBin Path to the jpegoptim binary
-     * @param bool   $stripAll     Strip all markers from output
-     * @param int    $max          Set maximum image quality factor
-     * @param bool   $progressive  Force output to be progressive
-     * @param string $tempDir      Directory where temporary file will be written
-     */
-    public function __construct(
-        $jpegoptimBin = '/usr/bin/jpegoptim',
-        $stripAll = true,
-        $max = null,
-        $progressive = true,
-        $tempDir = ''
-    ) {
-        $this->jpegoptimBin = $jpegoptimBin;
-        $this->stripAll = $stripAll;
-        $this->max = $max;
+        $this->strip = $strip;
+        $this->quality = $quality;
         $this->progressive = $progressive;
-        $this->tempDir = $tempDir ?: sys_get_temp_dir();
     }
 
     /**
-     * @param int $max
+     * @deprecated All post-processor setters have been deprecated in 2.2 for removal in 3.0. You must only use the
+     *             class's constructor to set the property state.
+     *
+     * @param int $maxQuality
      *
      * @return JpegOptimPostProcessor
      */
-    public function setMax($max)
+    public function setMax($maxQuality)
     {
-        $this->max = $max;
+        $this->triggerSetterMethodDeprecation(__METHOD__);
+        $this->quality = $maxQuality;
 
         return $this;
     }
 
     /**
+     * @deprecated All post-processor setters have been deprecated in 2.2 for removal in 3.0. You must only use the
+     *             class's constructor to set the property state.
+     *
      * @param bool $progressive
      *
      * @return JpegOptimPostProcessor
      */
     public function setProgressive($progressive)
     {
+        $this->triggerSetterMethodDeprecation(__METHOD__);
         $this->progressive = $progressive;
 
         return $this;
     }
 
     /**
-     * @param bool $stripAll
+     * @deprecated All post-processor setters have been deprecated in 2.2 for removal in 3.0. You must only use the
+     *             class's constructor to set the property state.
+     *
+     * @param bool $strip
      *
      * @return JpegOptimPostProcessor
      */
-    public function setStripAll($stripAll)
+    public function setStripAll($strip)
     {
-        $this->stripAll = $stripAll;
+        $this->triggerSetterMethodDeprecation(__METHOD__);
+        $this->strip = $strip;
 
         return $this;
     }
 
-    /**
-     * @param BinaryInterface $binary
-     * @param array           $options
-     *
+    /*
      * @throws ProcessFailedException
-     *
-     * @return BinaryInterface
      */
     public function process(BinaryInterface $binary, array $options = []): BinaryInterface
     {
-        $type = mb_strtolower($binary->getMimeType());
-        if (!in_array($type, ['image/jpeg', 'image/jpg'], true)) {
+        if (!$this->isBinaryTypeJpgImage($binary)) {
             return $binary;
         }
 
-        $tempDir = array_key_exists('temp_dir', $options) ? $options['temp_dir'] : $this->tempDir;
-        if (false === $input = tempnam($tempDir, 'imagine_jpegoptim')) {
-            throw new \RuntimeException(sprintf('Temp file can not be created in "%s".', $tempDir));
+        $file = $this->writeTemporaryFile($binary, $options, 'imagine-post-processor-jpegoptim');
+
+        $arguments = $this->getProcessArguments($options);
+        $arguments[] = $file;
+        $process = $this->createProcess($arguments, $options);
+        $process->run();
+
+        if (!$this->isSuccessfulProcess($process)) {
+            unlink($file);
+            throw new ProcessFailedException($process);
         }
 
-        $processArguments = [$this->jpegoptimBin];
+        $result = new Binary(file_get_contents($file), $binary->getMimeType(), $binary->getFormat());
 
-        $stripAll = array_key_exists('strip_all', $options) ? $options['strip_all'] : $this->stripAll;
-        if ($stripAll) {
-            $processArguments[] = '--strip-all';
-        }
-
-        $max = array_key_exists('max', $options) ? $options['max'] : $this->max;
-        if ($max) {
-            $processArguments[] = '--max='.$max;
-        }
-
-        $progressive = array_key_exists('progressive', $options) ? $options['progressive'] : $this->progressive;
-        if ($progressive) {
-            $processArguments[] = '--all-progressive';
-        } else {
-            $processArguments[] = '--all-normal';
-        }
-
-        $processArguments[] = $input;
-        if ($binary instanceof FileBinaryInterface) {
-            copy($binary->getPath(), $input);
-        } else {
-            file_put_contents($input, $binary->getContent());
-        }
-
-        $proc = new Process($processArguments);
-        $proc->run();
-
-        if (false !== mb_strpos($proc->getOutput(), 'ERROR') || 0 !== $proc->getExitCode()) {
-            unlink($input);
-            throw new ProcessFailedException($proc);
-        }
-
-        $result = new Binary(file_get_contents($input), $binary->getMimeType(), $binary->getFormat());
-
-        unlink($input);
+        unlink($file);
 
         return $result;
+    }
+
+    /**
+     * @param string[] $options
+     *
+     * @return string[]
+     */
+    private function getProcessArguments(array $options = []): array
+    {
+        $arguments = [$this->executablePath];
+
+        if ($options['strip_all'] ?? $this->strip) {
+            $arguments[] = '--strip-all';
+        }
+
+        if (isset($options['max'])) {
+            @trigger_error(sprintf('The "max" option was deprecated in 2.2 and will be removed in 3.0. '.
+                'Instead, use the "quality" option.'), E_USER_DEPRECATED);
+
+            if (isset($options['quality'])) {
+                throw new InvalidOptionException('the "max" and "quality" options cannot both be set', $options);
+            }
+
+            $options['quality'] = $options['max'];
+        }
+
+        if ($quality = $options['quality'] ?? $this->quality) {
+            if (!in_array($options['quality'], range(0, 100), true)) {
+                throw new InvalidOptionException('the "quality" option must be an int between 0 and 100', $options);
+            }
+
+            $arguments[] = sprintf('--max=%d', $quality);
+        }
+
+        if ($options['progressive'] ?? $this->progressive) {
+            $arguments[] = '--all-progressive';
+        } else {
+            $arguments[] = '--all-normal';
+        }
+
+        return $arguments;
     }
 }
