@@ -12,18 +12,12 @@
 namespace Liip\ImagineBundle\Imagine\Filter\PostProcessor;
 
 use Liip\ImagineBundle\Binary\BinaryInterface;
-use Liip\ImagineBundle\Binary\FileBinaryInterface;
+use Liip\ImagineBundle\Exception\Imagine\Filter\PostProcessor\InvalidOptionException;
 use Liip\ImagineBundle\Model\Binary;
 use Symfony\Component\Process\Exception\ProcessFailedException;
-use Symfony\Component\Process\Process;
 
-class OptiPngPostProcessor implements PostProcessorInterface
+class OptiPngPostProcessor extends AbstractPostProcessor
 {
-    /**
-     * @var string Path to optipng binary
-     */
-    protected $optipngBin;
-
     /**
      * If set --oN will be passed to optipng.
      *
@@ -36,83 +30,116 @@ class OptiPngPostProcessor implements PostProcessorInterface
      *
      * @var bool
      */
-    protected $stripAll;
+    protected $strip;
 
     /**
-     * Directory where temporary file will be written.
-     *
-     * @var string
+     * @param string $executablePath    Path to the optipng binary
+     * @param int    $level             Optimization level
+     * @param bool   $strip             Strip metadata objects
+     * @param string $temporaryRootPath Directory where temporary file will be written
      */
-    protected $tempDir;
-
-    /**
-     * Constructor.
-     *
-     * @param string $optipngBin Path to the optipng binary
-     * @param int    $level      Optimization level
-     * @param bool   $stripAll   Strip metadata objects
-     * @param string $tempDir    Directory where temporary file will be written
-     */
-    public function __construct($optipngBin = '/usr/bin/optipng', $level = 7, $stripAll = true, $tempDir = '')
+    public function __construct($executablePath = '/usr/bin/optipng', $level = 7, $strip = true, $temporaryRootPath = null)
     {
-        $this->optipngBin = $optipngBin;
+        parent::__construct($executablePath, $temporaryRootPath);
+
         $this->level = $level;
-        $this->stripAll = $stripAll;
-        $this->tempDir = $tempDir ?: sys_get_temp_dir();
+        $this->strip = $strip;
     }
 
-    /**
-     * @param BinaryInterface $binary
-     * @param array           $options
-     *
+    /*
      * @throws ProcessFailedException
-     *
-     * @return BinaryInterface|Binary
      */
     public function process(BinaryInterface $binary, array $options = []): BinaryInterface
     {
-        $type = mb_strtolower($binary->getMimeType());
-        if (!in_array($type, ['image/png'], true)) {
+        if (!$this->isBinaryTypePngImage($binary)) {
             return $binary;
         }
 
-        $tempDir = array_key_exists('temp_dir', $options) ? $options['temp_dir'] : $this->tempDir;
-        if (false === $input = tempnam($tempDir, 'imagine_optipng')) {
-            throw new \RuntimeException(sprintf('Temp file can not be created in "%s".', $tempDir));
+        $file = $this->writeTemporaryFile($binary, $options, 'imagine-post-processor-optipng');
+
+        $arguments = $this->getProcessArguments($options);
+        $arguments[] = $file;
+        $process = $this->createProcess($arguments, $options);
+        $process->run();
+
+        if (!$this->isSuccessfulProcess($process)) {
+            unlink($file);
+            throw new ProcessFailedException($process);
         }
 
-        $processArguments = [$this->optipngBin];
+        $result = new Binary(file_get_contents($file), $binary->getMimeType(), $binary->getFormat());
 
-        $level = array_key_exists('level', $options) ? $options['level'] : $this->level;
-        if (null !== $level) {
-            $processArguments[] = sprintf('--o%d', $level);
-        }
-
-        $stripAll = array_key_exists('strip_all', $options) ? $options['strip_all'] : $this->stripAll;
-        if ($stripAll) {
-            $processArguments[] = '--strip=all';
-        }
-
-        $processArguments[] = $input;
-
-        if ($binary instanceof FileBinaryInterface) {
-            copy($binary->getPath(), $input);
-        } else {
-            file_put_contents($input, $binary->getContent());
-        }
-
-        $proc = new Process($processArguments);
-        $proc->run();
-
-        if (false !== mb_strpos($proc->getOutput(), 'ERROR') || 0 !== $proc->getExitCode()) {
-            unlink($input);
-            throw new ProcessFailedException($proc);
-        }
-
-        $result = new Binary(file_get_contents($input), $binary->getMimeType(), $binary->getFormat());
-
-        unlink($input);
+        unlink($file);
 
         return $result;
+    }
+
+    /**
+     * @param string[] $options
+     *
+     * @return string[]
+     */
+    private function getProcessArguments(array $options = []): array
+    {
+        $arguments = [$this->executablePath];
+
+        if (null !== $level = $options['level'] ?? $this->level) {
+            if (!in_array($level, range(0, 7), true)) {
+                throw new InvalidOptionException('the "level" option must be an int between 0 and 7', $options);
+            }
+
+            $arguments[] = sprintf('-o%d', $level);
+        }
+
+        if (isset($options['strip_all'])) {
+            @trigger_error(sprintf('The "strip_all" option was deprecated in 2.2 and will be removed in 3.0. '.
+                'Instead, use the "strip" option.'), E_USER_DEPRECATED);
+
+            if (isset($options['strip'])) {
+                throw new InvalidOptionException('the "strip" and "strip_all" options cannot both be set', $options);
+            }
+
+            $options['strip'] = $options['strip_all'];
+        }
+
+        if ($strip = $options['strip'] ?? $this->strip) {
+            $arguments[] = '-strip';
+            $arguments[] = true === $strip ? 'all' : $strip;
+        }
+
+        if (isset($options['snip']) && true === $options['snip']) {
+            $arguments[] = '-snip';
+        }
+
+        if (isset($options['preserve_attributes']) && true === $options['preserve_attributes']) {
+            $arguments[] = '-preserve';
+        }
+
+        if (isset($options['interlace_type'])) {
+            if (!in_array($options['interlace_type'], range(0, 1), true)) {
+                throw new InvalidOptionException('the "interlace_type" option must be either 0 or 1', $options);
+            }
+
+            $arguments[] = '-i';
+            $arguments[] = $options['interlace_type'];
+        }
+
+        if (isset($options['no_bit_depth_reductions']) && true === $options['no_bit_depth_reductions']) {
+            $arguments[] = '-nb';
+        }
+
+        if (isset($options['no_color_type_reductions']) && true === $options['no_color_type_reductions']) {
+            $arguments[] = '-nc';
+        }
+
+        if (isset($options['no_palette_reductions']) && true === $options['no_palette_reductions']) {
+            $arguments[] = '-np';
+        }
+
+        if (isset($options['no_reductions']) && true === $options['no_reductions']) {
+            $arguments[] = '-nx';
+        }
+
+        return $arguments;
     }
 }
