@@ -12,9 +12,9 @@
 namespace Liip\ImagineBundle\Imagine\Filter\PostProcessor;
 
 use Liip\ImagineBundle\Binary\BinaryInterface;
+use Liip\ImagineBundle\Exception\Imagine\Filter\PostProcessor\InvalidOptionException;
 use Liip\ImagineBundle\Model\Binary;
 use Symfony\Component\Process\Exception\ProcessFailedException;
-use Symfony\Component\Process\Process;
 
 /**
  * pngquant post-processor, for optimal, web-safe, lossy png compression
@@ -25,76 +25,122 @@ use Symfony\Component\Process\Process;
  *
  * @author Alex Wilson <a@ax.gy>
  */
-class PngquantPostProcessor implements PostProcessorInterface
+class PngquantPostProcessor extends AbstractPostProcessor
 {
-    /**
-     * @var string Path to pngquant binary
-     */
-    protected $pngquantBin;
-
     /**
      * @var string Quality to pass to pngquant
      */
     protected $quality;
 
     /**
-     * Constructor.
-     *
-     * @param string $pngquantBin Path to the pngquant binary
-     * @param string $quality
+     * @param string $executablePath
+     * @param array  $quality
      */
-    public function __construct($pngquantBin = '/usr/bin/pngquant', $quality = '80-100')
+    public function __construct($executablePath = '/usr/bin/pngquant', $quality = [80, 100])
     {
-        $this->pngquantBin = $pngquantBin;
-        $this->setQuality($quality);
+        parent::__construct($executablePath);
+
+        $this->quality = $quality;
     }
 
     /**
+     * @deprecated All post-processor setters have been deprecated in 2.2 for removal in 3.0. You must only use the
+     *             class's constructor to set the property state.
+     *
      * @param string $quality
      *
      * @return PngquantPostProcessor
      */
     public function setQuality($quality)
     {
+        $this->triggerSetterMethodDeprecation(__METHOD__);
         $this->quality = $quality;
 
         return $this;
     }
 
-    /**
-     * @param BinaryInterface $binary
-     *
+    /*
      * @throws ProcessFailedException
-     *
-     * @return BinaryInterface
      */
     public function process(BinaryInterface $binary, array $options = []): BinaryInterface
     {
-        $type = mb_strtolower($binary->getMimeType());
-        if (!in_array($type, ['image/png'], true)) {
+        if (!$this->isBinaryTypePngImage($binary)) {
             return $binary;
         }
 
-        $processArguments = [$this->pngquantBin];
+        $arguments = $this->getProcessArguments($options);
+        $arguments[] = '-';
+        $process = $this->createProcess($arguments, $options);
+        $process->setInput($binary->getContent());
+        $process->run();
 
-        // Specify quality.
-        $tranformQuality = array_key_exists('quality', $options) ? $options['quality'] : $this->quality;
-        $processArguments[] = '--quality';
-        $processArguments[] = $tranformQuality;
-
-        // Read to/from stdout to save resources.
-        $processArguments[] = '-';
-        $proc = new Process($processArguments);
-        $proc->setInput($binary->getContent());
-        $proc->run();
-
-        // 98 and 99 are "quality too low" to compress current current image which, while isn't ideal, is not a failure
-        if (!in_array($proc->getExitCode(), [0, 98, 99], true)) {
-            throw new ProcessFailedException($proc);
+        if (!$this->isSuccessfulProcess($process, [0, 98, 99], [])) {
+            throw new ProcessFailedException($process);
         }
 
-        $result = new Binary($proc->getOutput(), $binary->getMimeType(), $binary->getFormat());
+        return new Binary($process->getOutput(), $binary->getMimeType(), $binary->getFormat());
+    }
 
-        return $result;
+    /**
+     * @param string[] $options
+     *
+     * @return string[]
+     */
+    private function getProcessArguments(array $options = []): array
+    {
+        $arguments = [$this->executablePath];
+
+        if ($quality = $options['quality'] ?? $this->quality) {
+            if (is_string($quality) && false !== strpos($quality, '-')) {
+                @trigger_error(sprintf('Passing the "quality" option as a string was deprecated in 2.2 and '.
+                    'will be removed in 3.0. Instead, pass wither an integer representing the max value or an array '.
+                    'representing the minimum and maximum values.'), E_USER_DEPRECATED);
+
+                $quality = array_map(function ($q) {
+                    return (int) $q;
+                }, explode('-', $quality));
+            }
+
+            if (!is_array($quality)) {
+                $quality = [0, (int) $quality];
+            }
+
+            if (1 === count($quality)) {
+                array_unshift($quality, 0);
+            }
+
+            if ($quality[0] > $quality[1]) {
+                throw new InvalidOptionException('the "quality" option cannot have a greater minimum value value than maximum quality value', $options);
+            }
+
+            if (!in_array($quality[0], range(0, 100), true) || !in_array($quality[1], range(0, 100), true)) {
+                throw new InvalidOptionException('the "quality" option value(s) must be an int between 0 and 100', $options);
+            }
+
+            $arguments[] = '--quality';
+            $arguments[] = sprintf('%d-%d', $quality[0], $quality[1]);
+        }
+
+        if (isset($options['speed'])) {
+            if (!in_array($options['speed'], range(1, 11), true)) {
+                throw new InvalidOptionException('the "speed" option must be an int between 1 and 11', $options);
+            }
+
+            $arguments[] = '--speed';
+            $arguments[] = $options['speed'];
+        }
+
+        if (isset($options['dithering'])) {
+            if (false === $options['dithering']) {
+                $arguments[] = '--nofs';
+            } elseif ($options['dithering'] >= 0 && $options['dithering'] <= 1) {
+                $arguments[] = '--floyd';
+                $arguments[] = $options['dithering'];
+            } elseif (true !== $options['dithering']) {
+                throw new InvalidOptionException('the "dithering" option must be a float between 0 and 1 or a bool', $options);
+            }
+        }
+
+        return $arguments;
     }
 }
