@@ -40,7 +40,7 @@ trait FilterTrait
      */
     public function filter($path, $filter, array $config = [], $resolver = null)
     {
-        return $this->cache->getBrowserPath($path, $filter, $config, $resolver);
+        return $this->cache->getBrowserPath(parse_url($path, PHP_URL_PATH), $filter, $config, $resolver);
     }
 
     /**
