@@ -75,7 +75,7 @@ class FileSystemInsecureLocatorTest extends AbstractFileSystemLocatorTest
         ];
 
         return array_map(function ($params) use ($prepend) {
-            return [[$prepend[mt_rand(0, count($prepend) - 1)], $params[0]], $params[1]];
+            return [[$prepend[mt_rand(0, \count($prepend) - 1)], $params[0]], $params[1]];
         }, static::provideLoadCases());
     }
 
