@@ -13,6 +13,7 @@ namespace Liip\ImagineBundle\Tests\Templating;
 
 use Liip\ImagineBundle\Imagine\Cache\CacheManager;
 use Liip\ImagineBundle\Templating\FilterExtension;
+use Twig\Extension\AbstractExtension;
 
 /**
  * @covers \Liip\ImagineBundle\Templating\FilterTrait
@@ -27,7 +28,7 @@ class FilterExtensionTest extends AbstractFilterTest
 
     public function testInstanceOfTwigFilter()
     {
-        $this->assertInstanceOf(\Twig_Extension::class, $this->createTemplatingMock());
+        $this->assertInstanceOf(AbstractExtension::class, $this->createTemplatingMock());
     }
 
     /**
@@ -37,7 +38,7 @@ class FilterExtensionTest extends AbstractFilterTest
      */
     protected function createTemplatingMock(CacheManager $manager = null)
     {
-        if (!class_exists(\Twig_Extension::class)) {
+        if (!class_exists(AbstractExtension::class)) {
             $this->markTestSkipped('Requires the twig/twig package.');
         }
 
