@@ -13,24 +13,9 @@ namespace Liip\ImagineBundle\Utility\Path;
 
 interface PathResolverInterface
 {
-    /**
-     * @param string $path
-     * @param string $filter
-     *
-     * @return string
-     */
-    public function getFilePath($path, $filter);
+    public function getFilePath(string $path, string $filter): string;
 
-    /**
-     * @param string $path
-     * @param string $filter
-     *
-     * @return string
-     */
-    public function getFileUrl($path, $filter);
+    public function getFileUrl(string $path, string $filter): string;
 
-    /**
-     * @return string
-     */
-    public function getCacheRoot();
+    public function getCacheRoot(): string;
 }
